exception

+++ JCAT_API/src/main/java/com/cg/jcat/api/controller/UserController.java
@@ -9,31 +9,52 @@ import com.cg.jcat.api.dao.UserModel;
 import com.cg.jcat.api.service.IUserService;
 
 @Component
-public class UserController implements IUserController{
-	
+public class UserController implements IUserController {
+
 	@Autowired
 	private IUserService userService;
 
 	@Override
 	public List<UserModel> getUsers() {
-		return userService.getUsers();
+		try {
+			return userService.getUsers();
+		} catch (Exception e) {
+
+			System.out.println("Error in getting users " + e);
+			return null;
+		}
 	}
 
 	@Override
-	public void saveUser(String createdBy, UserModel user) 
-	{
-		userService.saveUser(user,createdBy);
-		
+	public void saveUser(String createdBy, UserModel user) {
+		try {
+			if (createdBy != null && user != null) {
+				userService.saveUser(user, createdBy);
+			}
+		} catch (Exception e) {
+			System.out.print("Error while saving user" + e);
+		}
 	}
-
 
 	@Override
 	public void updateUserId(String modifiedBy, UserModel user) {
-		userService.updateUsers(user,modifiedBy);
+		try {
+			if(modifiedBy != null && user != null) {
+				userService.updateUsers(user, modifiedBy);
+			}
+		} catch (Exception e) {
+			System.out.print("Error while updating user" + e);
+		}
 	}
 
 	@Override
 	public void deleteById(int userId) {
-		userService.deleteById(userId);
+		try {
+			if (userId != 0) {
+				userService.deleteById(userId);
+			}
+		} catch (Exception e) {
+			System.out.print("Error while deleting user" + e);
+		}
 	}
 }
